WIP: refactor attribute class
